+++ packages/expect/src/index.js
@@ -335,19 +335,34 @@ const _validateResult = result => {
   }
 };
 
+function assertions(expected: number) {
+  const error = new Error();
+  if (Error.captureStackTrace) {
+    Error.captureStackTrace(error, assertions);
+  }
+
+  getState().expectedAssertionsNumber = expected;
+  getState().expectedAssertionsNumberError = error;
+}
+function hasAssertions(...args) {
+  const error = new Error();
+  if (Error.captureStackTrace) {
+    Error.captureStackTrace(error, hasAssertions);
+  }
+
+  utils.ensureNoExpected(args[0], '.hasAssertions');
+  getState().isExpectingAssertions = true;
+  getState().isExpectingAssertionsError = error;
+}
+
 // add default jest matchers
 setMatchers(matchers, true, expect);
 setMatchers(spyMatchers, true, expect);
 setMatchers(toThrowMatchers, true, expect);
 
 expect.addSnapshotSerializer = () => void 0;
-expect.assertions = (expected: number) => {
-  getState().expectedAssertionsNumber = expected;
-};
-expect.hasAssertions = (expected: any) => {
-  utils.ensureNoExpected(expected, '.hasAssertions');
-  getState().isExpectingAssertions = true;
-};
+expect.assertions = assertions;
+expect.hasAssertions = hasAssertions;
 expect.getState = getState;
 expect.setState = setState;
 expect.extractExpectedAssertionsErrors = extractExpectedAssertionsErrors;